+++ src/core/zowe/core_for_zowe_sdk/profile_manager.py
@@ -15,11 +15,13 @@ import os.path
 import warnings
 from copy import deepcopy
 from typing import Optional
+import logging
 
 import jsonschema
 from deepmerge import always_merger
 
 from .config_file import ConfigFile, Profile
+from .logger import Log
 from .credential_manager import CredentialManager
 from .custom_warnings import (
     ConfigNotFoundWarning,
@@ -57,10 +59,13 @@ class ProfileManager:
         self.project_config = ConfigFile(type=TEAM_CONFIG, name=appname)
         self.project_user_config = ConfigFile(type=USER_CONFIG, name=appname)
 
+        self.__logger = Log.registerLogger(__name__)
+
         self.global_config = ConfigFile(type=TEAM_CONFIG, name=GLOBAL_CONFIG_NAME)
         try:
             self.global_config.location = GLOBAL_CONFIG_LOCATION
         except Exception:
+            self.__logger.warning("Could not find Global Config Directory")
             warnings.warn(
                 "Could not find Global Config Directory, please provide one.",
                 ConfigNotFoundWarning,
@@ -70,6 +75,7 @@ class ProfileManager:
         try:
             self.global_user_config.location = GLOBAL_CONFIG_LOCATION
         except Exception:
+            self.__logger.warning("Could not find Global User Config Directory")
             warnings.warn(
                 "Could not find Global User Config Directory, please provide one.",
                 ConfigNotFoundWarning,
@@ -172,38 +178,48 @@ class ProfileManager:
             NamedTuple (data, name, secure_props_not_found)
         """
 
+        logger = logging.getLogger(__name__)
+
         cfg_profile = Profile()
         try:
             cfg_profile = cfg.get_profile(
                 profile_name=profile_name, profile_type=profile_type, validate_schema=validate_schema
             )
         except jsonschema.exceptions.ValidationError as exc:
+            logger.error(f"Instance was invalid under the provided $schema property, {exc}")
             raise jsonschema.exceptions.ValidationError(
                 f"Instance was invalid under the provided $schema property, {exc}"
             )
         except jsonschema.exceptions.SchemaError as exc:
+            logger.error(f"The provided schema is invalid, {exc}")
             raise jsonschema.exceptions.SchemaError(f"The provided schema is invalid, {exc}")
         except jsonschema.exceptions.UndefinedTypeCheck as exc:
+            logger.error(f"A type checker was asked to check a type it did not have registered, {exc}")
             raise jsonschema.exceptions.UndefinedTypeCheck(
                 f"A type checker was asked to check a type it did not have registered, {exc}"
             )
         except jsonschema.exceptions.UnknownType as exc:
+            logger.error(f"Unknown type is found in schema_json, {exc}")
             raise jsonschema.exceptions.UnknownType(f"Unknown type is found in schema_json, exc")
         except jsonschema.exceptions.FormatError as exc:
+            logger.error(f"Validating a format config_json failed for schema_json, {exc}")
             raise jsonschema.exceptions.FormatError(f"Validating a format config_json failed for schema_json, {exc}")
         except ProfileNotFound:
             if profile_name:
+                logger.warning(f"Profile '{profile_name}' not found in file '{cfg.filename}'")
                 warnings.warn(
                     f"Profile '{profile_name}' not found in file '{cfg.filename}', returning empty profile instead.",
                     ProfileNotFoundWarning,
                 )
             else:
+                logger.warning(f"Profile of type '{profile_type}' not found in file '{cfg.filename}'")
                 warnings.warn(
                     f"Profile of type '{profile_type}' not found in file '{cfg.filename}', returning empty profile"
                     f" instead.",
                     ProfileNotFoundWarning,
                 )
         except Exception as exc:
+            logger.warning(f"Could not load '{cfg.filename}' at '{cfg.filepath}'" f"because {type(exc).__name__}'{exc}'")
             warnings.warn(
                 f"Could not load '{cfg.filename}' at '{cfg.filepath}'" f"because {type(exc).__name__}'{exc}'.",
                 ConfigNotFoundWarning,
@@ -218,6 +234,7 @@ class ProfileManager:
         check_missing_props: bool = True,
         validate_schema: Optional[bool] = True,
         override_with_env: Optional[bool] = False,
+        suppress_config_file_warnings: Optional[bool] = True,
     ) -> dict:
         """Load connection details from a team config profile.
         Returns
@@ -236,7 +253,9 @@ class ProfileManager:
         If `profile_type` is not base, then we will load properties from both
         `profile_type` and base profiles and merge them together.
         """
+        
         if profile_name is None and profile_type is None:
+            self.__logger.error(f"Failed to load profile as both profile_name and profile_type are not set")
             raise ProfileNotFound(
                 profile_name=profile_name,
                 error_msg="Could not find profile as both profile_name and profile_type is not set.",
@@ -254,12 +273,13 @@ class ProfileManager:
         cfg_name = None
         cfg_schema = None
         cfg_schema_dir = None
-
+        
         for cfg_layer in (self.project_user_config, self.project_config, self.global_user_config, self.global_config):
             if cfg_layer.profiles is None:
                 try:
-                    cfg_layer.init_from_file(validate_schema)
+                    cfg_layer.init_from_file(validate_schema, suppress_config_file_warnings)
                 except SecureProfileLoadFailed:
+                    self.__logger.warning(f"Could not load secure properties for {cfg_layer.filepath}")
                     warnings.warn(
                         f"Could not load secure properties for {cfg_layer.filepath}",
                         SecurePropsNotFoundWarning,
@@ -314,6 +334,7 @@ class ProfileManager:
                     missing_props.add(item)
 
             if len(missing_props) > 0:
+                self.__logger.error(f"Failed to load secure values: {missing_props}")
                 raise SecureValuesNotFound(values=missing_props)
 
         warnings.resetwarnings()
@@ -366,6 +387,7 @@ class ProfileManager:
                 highest_layer = layer
 
         if highest_layer is None:
+            self.__logger.error(f"Could not find a valid layer for {json_path}")
             raise FileNotFoundError(f"Could not find a valid layer for {json_path}")
 
         return highest_layer